requirements.txt
Flask==2.0.1
flask-restx==0.5.1
PyJWT==2.3.0
nvidia-ml-py3==7.352.0
tenacity==8.0.1
Werkzeug==2.0.1
Flask-Limiter==2.6.0
pyOpenSSL==20.0.1
